refactor: Refactor Pill to TS and better typings for codebase
