+++ packages/forma-36-react-components/src/components/Pill/index.ts
