+++ packages/forma-36-react-components/src/components/Spinner/Spinner.tsx
@@ -16,7 +16,7 @@ interface SpinnerProps {
 }
 
 class Spinner extends React.Component<SpinnerProps> {
-  static defaultProps: Partial<SpinnerProps> = {
+  static defaultProps = {
     extraClassNames: undefined,
     testId: 'cf-ui-spinner',
     size: 'default',
