+++ packages/forma-36-react-components/tools/plop-templates/components/component.tsx.hbs
@@ -11,7 +11,7 @@ export interface {{pascalCase name}}Props {
 
 class {{pascalCase name}} extends React.Component<{{pascalCase name}}Props> {
 
-  static defaultProps: Partial<{{pascalCase name}}Props> = {
+  static defaultProps = {
     extraClassNames: undefined,
     testId: 'cf-ui-{{kebabCase name}}',
   };
