+++ packages/forma-36-react-components/typings.d.ts
@@ -1,4 +1,4 @@
 declare module '*.md' {
-  const content: string;
-  export default content;
+  const value: string;
+  export default value;
 }
